* generic/tclThreadTest.c (Tcl_ThreadObjCmd): cast fixes

+++ generic/tclThreadTest.c
@@ -11,7 +11,7 @@
  * See the file "license.terms" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * RCS: @(#) $Id: tclThreadTest.c,v 1.10 2000/05/02 22:02:36 kupries Exp $
+ * RCS: @(#) $Id: tclThreadTest.c,v 1.11 2001/03/31 07:55:50 hobbs Exp $
  */
 
 #include "tclInt.h"
@@ -252,7 +252,7 @@ Tcl_ThreadObjCmd(dummy, interp, objc, objv)
 
 		if ((len > 1) &&
 		    (script [0] == '-') && (script [1] == 'j') &&
-		    (0 == strncmp (script, "-joinable", len))) {
+		    (0 == strncmp (script, "-joinable", (size_t) len))) {
 		    joinable = 1;
 		    script   = "testthread wait"; /* Just enter the event loop
 						   */
@@ -270,7 +270,7 @@ Tcl_ThreadObjCmd(dummy, interp, objc, objv)
 
 		joinable = ((len > 1) &&
 			    (script [0] == '-') && (script [1] == 'j') &&
-			    (0 == strncmp (script, "-joinable", len)));
+			    (0 == strncmp (script, "-joinable", (size_t) len)));
 
 		script = Tcl_GetString(objv[3]);
 	    } else {
